Rename LLAMAS -> fault_sdk
