--- libraries/LLAMAS/LLAMAS.ino
@@ -1,6 +1,0 @@
-
-void setup() {
-}
-
-void loop() {
-}
