+++ libraries/fault_sdk/faultManagement.c
